MOIN BON

+++ src/main/java/com/dcspa/prism/service/NiveauAlphaService.java
@@ -20,12 +20,12 @@ public class NiveauAlphaService {
 
 	@Transactional(readOnly = true)
 	public Optional<NiveauAlpha> findById(Integer id) {
-		return id == null ? Optional.empty() : repository.findById(id.longValue());
+		return id == null ? Optional.empty() : repository.findById(id);
 	}
 
 	@Transactional
 	public NiveauAlpha save(NiveauAlpha entity) { return repository.save(entity); }
 
 	@Transactional
-	public void deleteById(Integer id) { if (id != null) repository.deleteById(id.longValue()); }
+	public void deleteById(Integer id) { if (id != null) repository.deleteById(id); }
 }
